requirements.txt
langchain
langchain_openai
langchain_experimental
llama-index
python-dotenv
markdown-pdf
google-generativeai
beautifulsoup4